guest: Use systemd
Simplify uml-init.sh and make it compatible with the host configurations
handled by systemd.

Using a proper init system fixes the session leader, which wasn't really
an issue for our case, but it is cleaner. It is still possible to
launch an interactive shell, which is useful for debugging.

Using this systemd configuration slow down the boot time, but it should
still be quick enough: from 0.2 to 0.5 second.

Thanks to the SYSTEMD_UNIT_PATH kernel boot parameter, we don't rely on
the host (root-owned) files, and we can leverage the PATH environement
variable thanks to direct call to the host's Bash.

Use /run as TMPDIR because it is, by convention, a tmpfs mount point
that will not reflect on the host.

Signed-off-by: Mickaël Salaün <mic@digikod.net>