test-rust: Fix CI's TTY issue
The GitHub CI is doing something special with /dev/console, which
creates issues with a second layer of emulation (i.e. UML).

Use pipes as stdio instead of the CI's TTY to avoid missing some
stdout/stderr. We can still use run-uml.sh on a workstation to create an
interactive shell with a proper TTY.

Add a timeout to cat because when using a pipe, all the connected peers
must exit.

Signed-off-by: Mickaël Salaün <mic@digikod.net>