check-linux: Build old kselftest with UML
If we try to install kernel headers with UML we get: "Headers not
exportable for UML."

We can use the x86_64 architecture instead, but this doesn't work if the
headers_install target is a dependency of another one (e.g. Kselftest's
install on Linux v5.15).

Instead, quickly fix the Kselftest's Makefile to not call
headers_install, which is explicitly called before.

Make unpatch_all() generic because only one trap can be registered at a
time, and use $PATCHES to save the applied fixes.

Signed-off-by: Mickaël Salaün <mic@digikod.net>