uml: Mount tmpfs on /tmp
Create new tmpfs mount points for /tmp and set the befault value of
$TMPDIR accordingly.  This is cleaner and it enables to pass the
fs_test:ruleset_fd_transfer test on kernels without hostfs support.

Create the return value file shared with uml-init.sh in the kernel's
directory.  This enables to not rely on $TMPDIR nor /tmp, and it is
simpler to collocate this file with the kernel being tested.

Signed-off-by: Mickaël Salaün <mic@digikod.net>