check-linux: Build the sandboxer sample on UML
The kernel configuration should require CONFIG_HEADERS_INSTALL=y but it
is not compatible with UML.  Patch Landlock sample's dependencies to
make it possible to build it (if the headers are installed).

Don't remove the sample once built because check_build might not be
compatible with UML.

Check that the sample is always built (with the "build" command).

Clean up patches.

Signed-off-by: Mickaël Salaün <mic@digikod.net>