check-linux: Build with runtime checks
Build the kernel with useful runtime check features, except for the
build_light command.  This increases the size of Linux v6.10.8 image
from 7M to 105M, the build directory size from 104M to 630M, and the
build time from 1m24 to 2m27.

Force panic on kernel issues, and double check with the kselftest
command that no kernel error is detected during a test.

Increase available memory to avoid kernel issues.

Signed-off-by: Mickaël Salaün <mic@digikod.net>