uml-init: remove host systemd dependency
The current UML workflow depends on the host system init being systemd:
- boot UML with host filesystem and the default init=/sbin/init
- otherwise, systemd parses the SYSTEMD_UNIT_PATH from kernel
  commandline and starts landlock-test.service
- landlock-test.service starts bash uml-init.sh
- uml-init.sh extracts commandline from kernel's /proc/cmdline (the
  part after --) and runs it

This dependency makes it complicated to run tests in a container, or on
OSes with a different init system.

Replace the init with:
`/bin/bash /path/to/uml-init.sh <user-provided-commandline>`

This only changes the commandline interface of `uml-init.sh`, not that
of `uml-run.sh`. The only change required in the userspace setup is to
mount /proc then link /dev/fd to /proc/self/fd/ for bash redirections
to work.

There's just a little tweak to pass the commandline: a first -- is
needed to tell the uml-run.sh script the split between
{kernel+environment variables} and {usermode commandline}. Now a
different -- is needed to tell the kernel the split between its argv[0]
(/bin/bash) and argv[1..] (usermode commandline)

With this patch, dependencies are essentially bash + iproute2 (to set
the loopback interface up in network tests), which makes it possible
to run tests in a container environment without systemd.

Signed-off-by: Matthieu Buffet <matthieu@buffet.re>